requirements.txt
Django==5.1.5
celery==5.4.0
redis==5.2.1
flower==2.0.1
selenium>=4.15.0
selenium-wire>=5.1.0
webdriver-manager>=4.0.1
blinker==1.7.0
setuptools==75.8.0
python-dotenv>=1.0.0